main_service/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
flask-jwt-extended>=4.7.1
gunicorn==23.0.0
mysqlclient==2.2.6
python-dotenv==1.0.1
Flask-SocketIO==5.4.1
Flask-CORS==5.0.0
pydantic==2.10.3
requests==2.32.3